requirements.txt
streamlit
Pillow
qrcode
opencv-python-headless
numpy
